Extended Google Analytics settings with all possible paramenters.
Removed the concept of a single IEventsProcessor, and instead refactored the
code so you can have multiple processors, but each one can only be registered
once. This was necessary if you want to combine ie. Google Analytics campaign
tracking and SendGrid.

Refactoring of Type-discoveries to use MEF instead of simple Assembly scanning.
This means that the Export attribute is needed to export IMailTemplatesProvider
and IMailClient implementations.

Bumbed version to 0.6.0.